// generated/src/app/api-models/cap-age-band-detail.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapAgeBandDetail {
  ageBandId: string;
  seqAgeBandDtl: number;
  memberGender: string;
  ageFrom: number;
  ageThru: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}